+++ changelog.md
@@ -1,3 +1,10 @@
+## v1.0.1
+
+- Fixed the artist logo bug (NCS and Chompo) not aligned with the song name
+- Replace all of the vanilla nodes to hide instead of removing it entirely
+- Tweaked the comments date label
+- Fixed bug where user with creator points doesn't show their rated level
+
 ## v1.0.0
 
 - get this profile page design out of my head and into the game